moved the print board functionality to tic tac toe - encapsulation
